final changes, fixing erros...

+++ backend/Popcorn_Prophet/src/main/java/com/popcorn_prophet/popcorn_prophet/service/ProductService.java
@@ -32,6 +32,9 @@ public class ProductService {
             if (productToDelete.get().getImage() != null) {
                 imageService.deleteImage(productToDelete.get().getImage().getId());
             }
+            if(cartService.removeItemFromAllCarts(id).isEmpty()){
+                return Optional.empty();
+            }
             productRepository.deleteById(id);
 
             cartService.recalculateCartsTotal();
